+++ adapters/zapadapter/logger.go
@@ -2,6 +2,7 @@ package zapadapter
 
 import (
 	"github.com/goph/logur"
+	"github.com/goph/logur/internal/keyvals"
 	"go.uber.org/zap"
 )
 
@@ -19,41 +20,23 @@ func New(logger *zap.SugaredLogger) logur.Logger {
 	return &adapter{logger}
 }
 
-func (a *adapter) Trace(msg string) {
+func (a *adapter) Trace(msg string, fields map[string]interface{}) {
 	// Fall back to Debug
-	a.Debug(msg)
+	a.Debug(msg, fields)
 }
 
-func (a *adapter) Debug(msg string) {
-	a.logger.Debug(msg)
+func (a *adapter) Debug(msg string, fields map[string]interface{}) {
+	a.logger.Debugw(msg, keyvals.FromMap(fields)...)
 }
 
-func (a *adapter) Info(msg string) {
-	a.logger.Info(msg)
+func (a *adapter) Info(msg string, fields map[string]interface{}) {
+	a.logger.Infow(msg, keyvals.FromMap(fields)...)
 }
 
-func (a *adapter) Warn(msg string) {
-	a.logger.Warn(msg)
+func (a *adapter) Warn(msg string, fields map[string]interface{}) {
+	a.logger.Warnw(msg, keyvals.FromMap(fields)...)
 }
 
-func (a *adapter) Error(msg string) {
-	a.logger.Error(msg)
-}
-
-func (a *adapter) WithFields(fields map[string]interface{}) logur.Logger {
-	keyvals := make([]interface{}, len(fields)*2)
-	i := 0
-
-	for key, value := range fields {
-		keyvals[i] = key
-		keyvals[i+1] = value
-
-		i += 2
-	}
-
-	if keyvals == nil {
-		return a
-	}
-
-	return &adapter{a.logger.With(keyvals...)}
+func (a *adapter) Error(msg string, fields map[string]interface{}) {
+	a.logger.Errorw(msg, keyvals.FromMap(fields)...)
 }