+++ adapters/zerologadapter/logger.go
@@ -15,27 +15,23 @@ func New(logger zerolog.Logger) logur.Logger {
 	return &adapter{logger}
 }
 
-func (a *adapter) Trace(msg string) {
+func (a *adapter) Trace(msg string, fields map[string]interface{}) {
 	// Fall back to Debug
-	a.Debug(msg)
+	a.Debug(msg, fields)
 }
 
-func (a *adapter) Debug(msg string) {
-	a.logger.Debug().Msg(msg)
+func (a *adapter) Debug(msg string, fields map[string]interface{}) {
+	a.logger.Debug().Fields(fields).Msg(msg)
 }
 
-func (a *adapter) Info(msg string) {
-	a.logger.Info().Msg(msg)
+func (a *adapter) Info(msg string, fields map[string]interface{}) {
+	a.logger.Info().Fields(fields).Msg(msg)
 }
 
-func (a *adapter) Warn(msg string) {
-	a.logger.Warn().Msg(msg)
+func (a *adapter) Warn(msg string, fields map[string]interface{}) {
+	a.logger.Warn().Fields(fields).Msg(msg)
 }
 
-func (a *adapter) Error(msg string) {
-	a.logger.Error().Msg(msg)
-}
-
-func (a *adapter) WithFields(fields map[string]interface{}) logur.Logger {
-	return &adapter{a.logger.With().Fields(fields).Logger()}
+func (a *adapter) Error(msg string, fields map[string]interface{}) {
+	a.logger.Error().Fields(fields).Msg(msg)
 }