+++ integrations/invisionlog/logger.go
@@ -19,19 +19,19 @@ func New(l logur.Logger) log.Logger {
 }
 
 func (l *logger) Debug(msg ...interface{}) {
-	l.logger.Debug(fmt.Sprint(msg...))
+	l.logger.Debug(fmt.Sprint(msg...), nil)
 }
 
 func (l *logger) Info(msg ...interface{}) {
-	l.logger.Info(fmt.Sprint(msg...))
+	l.logger.Info(fmt.Sprint(msg...), nil)
 }
 
 func (l *logger) Warn(msg ...interface{}) {
-	l.logger.Warn(fmt.Sprint(msg...))
+	l.logger.Warn(fmt.Sprint(msg...), nil)
 }
 
 func (l *logger) Error(msg ...interface{}) {
-	l.logger.Error(fmt.Sprint(msg...))
+	l.logger.Error(fmt.Sprint(msg...), nil)
 }
 
 func (l *logger) Debugln(msg ...interface{}) {
@@ -68,5 +68,5 @@ func (l *logger) Errorf(format string, args ...interface{}) {
 
 // WithFields returns a new logger with the additional supplied fields.
 func (l *logger) WithFields(fields log.Fields) log.Logger {
-	return &logger{l.logger.WithFields(logur.Fields(fields))}
+	return &logger{logur.WithFields(l.logger, logur.Fields(fields))}
 }