+++ logger.go
@@ -9,33 +9,30 @@ type Logger interface {
 	//
 	// Even more fine-grained information than Debug events.
 	// Loggers not supporting this level should fall back to Debug.
-	Trace(msg string)
+	Trace(msg string, fields map[string]interface{})
 
 	// Debug logs a Debug event.
 	//
 	// A verbose series of information events.
 	// They are useful when debugging the system.
-	Debug(msg string)
+	Debug(msg string, fields map[string]interface{})
 
 	// Info logs an Info event.
 	//
 	// General information about what's happening inside the system.
-	Info(msg string)
+	Info(msg string, fields map[string]interface{})
 
 	// Warn logs a Warn(ing) event.
 	//
 	// Non-critical events that should be looked at.
-	Warn(msg string)
+	Warn(msg string, fields map[string]interface{})
 
 	// Error logs an Error event.
 	//
 	// Critical events that require immediate attention.
 	// Loggers commonly provide Fatal and Panic levels above Error level,
 	// but exiting and panicing is out of scope for a logging library.
-	Error(msg string)
-
-	// WithFields appends structured fields to a new (child) logger instance.
-	WithFields(fields map[string]interface{}) Logger
+	Error(msg string, fields map[string]interface{})
 }
 
 // Fields is used to define structured fields which are appended to log events.
@@ -43,4 +40,4 @@ type Logger interface {
 type Fields map[string]interface{}
 
 // LogFunc is a function recording a log event.
-type LogFunc func(msg string)
+type LogFunc func(msg string, fields map[string]interface{})